+++ CHANGES.md
@@ -21,6 +21,7 @@
 - Change contacts filter to rename the anachronistic `Online` state to `Available`.
 - Enable [reuse_scram_keys](https://conversejs.org/docs/html/configuration.html#reuse-scram-keys) by default.
 - New loadEmojis hook, to customize emojis at runtime.
+- Fix custom emojis path when assets_path is not the default path.
 
 ### Breaking changes: